Move `CreateBoxCondition` command
