+++ wcfsetup/install/files/lib/system/box/BoxHandler.class.php
@@ -4,7 +4,7 @@ namespace wcf\system\box;
 
 use wcf\data\box\Box;
 use wcf\data\box\BoxList;
-use wcf\system\box\command\CreateBoxCondition;
+use wcf\command\box\CreateBoxCondition;
 use wcf\system\box\command\CreateBoxToPageAssignments;
 use wcf\system\event\EventHandler;
 use wcf\system\request\RequestHandler;
